FUSETOOLS2-1266 - move a dedicated folder the requirements code
This is the code mostly duplicated from VS Code Java which is checking
that a compatible JRE is available.
